add: populate the database on the first run for testing.

+++ app/src/main/java/com/james/jkomarket/product/model/Listing.kt
@@ -22,7 +22,7 @@ data class Listing(
 
     var user_name: String,
 
-    var category: Int,
+    var category: Long,
 
     @ColumnInfo(name = "created_at")
     var createAt: Long = Date(System.currentTimeMillis()).time,
